disk_block_cache: Restore file naming so the diff is easier.
